+++ CollaboratedHRM/Setup/Base.cs
@@ -1,11 +1,6 @@
 ﻿using BoDi;
 using OpenQA.Selenium.Chrome;
 using OpenQA.Selenium;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using TechTalk.SpecFlow;
 
 namespace CollaboratedHRM.Setup
