+++ CollaboratedHRM/StepDefinition/LoginStepDefinition.cs
@@ -1,6 +1,5 @@
 using CollaboratedHRM.Pages;
 using CollaboratedHRM.Setup;
-using System;
 using TechTalk.SpecFlow;
 
 namespace CollaboratedHRM.StepDefinition
@@ -45,12 +44,8 @@ namespace CollaboratedHRM.StepDefinition
         [Then(@"Then user is logged in successfully")]
         public void ThenThenUserIsLoggedInSuccessfully()
         {
-            var Getbarchart = _homePage.BarchartDisplay();
-            Assert.IsTrue(Getbarchart);
-            var DisplayBarchart = _homePage.DisplayDashbord();
-            Assert.IsTrue(DisplayBarchart);
-
-
+            Assert.IsTrue(_homePage.IsBarchartDisplay());
+            Assert.IsTrue(_homePage.IsConfirmationPieChartDisplayed());
         }
     }
 }
